Extract webpacker:compile into its own job
Having this be a `run` step in the system_specs job worked fine but
didn't feel congruent with the other modular pieces of our build
config.

It also makes it more difficult to identify issues with the webpacker
command, eg noticing if it takes too long, and make it smarter, eg with
caching.