Cache webpacker:compile artifacts in CI build
I'm not sure if there's a useful checksum to append to the cache key on
this. Moreover, i believe webpacker comes with smarts to figure out what
needs to be recompiled based on checksums it stores in
tmp/cache/webpacker, so even a somewhat stale cache hit might be ok. For
now, setting up a key based on the branch name, then defaulting to main.